Refactor Step and NewStep to only take single actor, fix status
