--- server/@types/NewStepType.ts
@@ -1,10 +1,0 @@
-export type NewStep = {
-  description?: string
-  actor: Actor[]
-  status?: string
-}
-
-type Actor = {
-  actor: string
-  actorOptionId: number
-}
