+++ server/@types/StepType.ts
@@ -1,12 +1,20 @@
-export type Step = {
-  uuid: string
-  description: string
-  status?: string
-  creationDate?: string
-  actor: Actor[]
+export enum StepStatus {
+  NOT_STARTED = 'NOT_STARTED',
+  IN_PROGRESS = 'IN_PROGRESS',
+  COMPLETED_ = 'COMPLETED',
+  CANNOT_BE_STARTED_YET = 'CANNOT_BE_STARTED_YET',
 }
 
-type Actor = {
+export type NewStep = {
+  description: string
+  status: StepStatus
   actor: string
-  actorOptionId: number
+}
+
+export type Step = {
+  uuid: string
+  creationDate: string
+  description: string
+  status: StepStatus
+  actor: string
 }
