+++ server/routes/steps/StepsController.ts
@@ -1,11 +1,11 @@
-import { Request, Response, NextFunction } from 'express'
+import { NextFunction, Request, Response } from 'express'
 import StepService from '../../services/sentence-plan/stepsService'
 import locale from './locale.json'
 import { FORMS } from '../../services/formStorageService'
-import { NewStep } from '../../@types/NewStepType'
 import URLs from '../URLs'
 import options from './options'
 import { toKebabCase } from '../../utils/utils'
+import { NewStep, StepStatus } from '../../@types/StepType'
 
 export default class StepsController {
   constructor(private readonly stepService: StepService) {}
@@ -45,20 +45,11 @@ export default class StepsController {
       const actorArray = Array.isArray(actor) ? actor : [actor]
       const actorNumbers: number[] = actorArray.map((item: string | number): number => Number(item))
       const mappedActor: Array<any> = payloadOptions.filter(option => actorNumbers.includes(option.value))
-      const actors = transformActor(mappedActor)
-      const payload = { description: stepName, status: '', actor: actors } as NewStep
+      const payload = { description: stepName, status: StepStatus.NOT_STARTED, actor: mappedActor[0].text } as NewStep
       await this.stepService.saveSteps([payload], currentGoal)
       res.redirect(`${URLs.PLAN_SUMMARY}?status=success`)
     } catch (e) {
       next(e)
     }
   }
-}
-
-function transformActor(mappedActor: Array<any>): Array<any> {
-  const actorArray: Array<any> = []
-  mappedActor.forEach(obj => {
-    actorArray.push({ actor: obj.text, actorOptionId: obj.value })
-  })
-  return actorArray
 }
