+++ server/testutils/data/stepData.ts
@@ -1,16 +1,10 @@
-import { NewStep } from '../../@types/NewStepType'
-import { Step } from '../../@types/StepType'
+import { NewStep, Step, StepStatus } from '../../@types/StepType'
 import { testGoal } from './goalData'
 
 export const testNewStep: NewStep = {
   description: 'A test step',
-  actor: [
-    {
-      actor: 'Test actor',
-      actorOptionId: 1,
-    },
-  ],
-  status: 'PENDING',
+  actor: 'Test actor',
+  status: StepStatus.NOT_STARTED,
 }
 export const testStep: Step = {
   id: 123,
